Manual: Shorten the introduction down to 1 page
Move references to appendices.
Condense the information a bit.
Add a short aircraft specification table.

Move 'Generic FlightGear Operations' to the 'Operation' part.

Remove the FG settings recommandations: as far as I can tell, these were
rather generic FG recommandations, and not something that could break
compatibility with the Viggen. Thus I think it is not appropriate in
this manual.